+++ src/examples/global-store/game/game.page.jsx
