+++ src/examples/global-store/game/games.json
